+++ src/components/tooltip/TooltipNote.tsx
@@ -14,7 +14,7 @@ const Note = styled.div`
   white-space: nowrap;
   pointer-events: none;
   z-index: -1;
-  background: rgba(0, 0, 0, 0.4);
+  background: rgba(0, 0, 0, 0.6);
   margin: 6px;
 `;